requirements.txt
accelerate==1.4.0
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.8.0
audioop-lts==0.2.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
einops==0.8.1
fastapi==0.115.11
ffmpy==0.5.0
filelock==3.17.0
fsspec==2025.3.0
gradio==5.20.1
gradio_client==1.7.2
gradio_imageslider==0.0.20
groovy==0.1.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
kornia==0.8.0
kornia_rs==0.1.8
lazy_loader==0.4
loadimg==0.4.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
opencv-python==4.11.0.86
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==11.1.0
prettytable==3.15.1
psutil==5.9.8
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
ruff==0.9.10
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.2
semantic-version==2.10.0
setuptools==76.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
spaces==0.32.0
starlette==0.46.1
sympy==1.13.1
tifffile==2025.2.18
timm==1.0.15
tokenizers==0.21.0
tomlkit==0.13.2
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.49.0
typer==0.15.2
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
websockets==15.0.1
